// src/app/Models/Entities/Client.ts
export class Client {
    id: any;
    nom: string;
    reference: string;
    telephone: string;
    fax: string;
    email: string;
    siteWeb: string;
    siret: string;
    tvaIntraCommunautaire: string;
    codeComptable: string;
    adresses: string;
    memos: string;
    contacts: string;
    historique: string;
    idAgent;
    idGroupe;
    codeclient: string; 
}